Carpetas Services, Repository, Controllers
Puse las propiedades que muestra en el video en el arch de configuración, cree repositories y services de la clase Categorias. Creé el servicio Base del que extienden los otros servicios

+++ src/main/java/Entities/Categoria.java
@@ -18,7 +18,7 @@ public class Categoria {
     @Id
     @Column(name = "idcategoria")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int idCategoria;
+    private Long idCategoria;
 
     @Column(name = "nombre")
     private String nombre;
